Rename Kernel/init.m -> Kernel/init.wl
The extension of the package initialization file is changed to match
the extension of the package file (RunningCouplings.wl) in order to
comply with the Wolfram Language's package loading conventions.
Kernel/init.wl is now sourced when the package is loaded with Needs or
Get, instead of RunningCouplings.wl directly.