Fix bug in MSSM renormalization with Dirac neutrinos
A typo in the options specified for the MSSM NDSolve algorithm causes
the renormalization to fail in the Dirac neutrino case. This revision
removes the block of code. This revision removes the problematic
options specification entirely, as it should add nothing in the Dirac
case and is functionally inert in the Majorana case as well.