Fix bug causing SM -> MSSM threshold uncertainty estimate to fail
Improper data type (`Around`) in the utility function `toRoundedNumber`
was causing the calculation of threshold correction uncertainty in the
internal fuction `uRunningCouplings` to silently fail (returning an
uncertianty estimate of zero). This is corrected by extracting the
`"Value"` before passing the object into `toRoundedNumber`.

Another solution is to adjust the definition of `toRoundedNumber` to do
this automatically for `Around` types, but it is not obvious that this
will always be the intended behavior. Better would be to add a warning
message if `toRoundedNumber` (and other functions) encounters an
improper data type.