Fix bugs in initial SMDR compilation and running
- Fix syntax errors (extra commas introduced in bugfix for issue #3)
causing failure of function `BuildSMDR`.
- Reorganize and clean up `BuildSMDR` function.
- Fix bug in function `RunningCouplings`Private`roundTo` causing
incorrect uncertainty (or no uncertainty) to be interpreted during file
import in function RunningCouplings`Private`smdrInputParameters.
- Fix bug in application of function
`RunningCouplings`Private`uRunningcouplings` to model SMDR at the scale
`"TopQuarkMassScale"` (this is the initial extraction of the SM
couplings) causing the calculated uncertainty to be dropped in the
returned couplings due to a mismatch between the keys of the input
coupling association and the output coupling association. The option
`"UncertainCouplings"` is added to `uRunningcouplings` to allow direct
specification of the list of couplings which should be varied in the
uncertainty analysis and for which uncertainty should be calculated.
This option is now explicitly supplied to all calls to
`uRunningcouplings` for SMDR at the `"TopQuarkMassScale"`.